make version of totals which takes array directly
association page can update that async based on assignment callbacks
